do not export secp256k1 ctypes handle directly
exporting ctypes handle directly creates incentive to use it for
different purposes outside the library. While this might be useful,
it is also dangerous if not done with the same care that is needed
when programming in C. Correct way is to export the function that will
load the library, initialize the default function definitions and
return the new handle to those who are confident enough to use it.